requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.23.0
pillow>=9.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6